Replace ContainerSpec.GraphCaching with ContainerTests.GraphCaching.
